Working on solid trend alts
Overlap is computing and can be rendering in oz,
but the underlying vega plot lacks the svg
defs when we go to emit it.

We might be able to get around this by rendering to
svg, then prepending the pattern defs into the
svg header as a post process.

For now, we have the fallback option to use
a simple colored trend for overlap.