Lots of compatiblity fixes, finishing nlist
Ran into a problem with writing excel workbooks
that led down a POI rabbit hole. Specifically,
the batik wrapper was pulling in some xml
lib called xerces that is notorious for messing
with xml dependencies (as with docjure, and poi).

Ended up fixing it by pushing exclusions all over
the batik wrapper, pulling in current batik-all,
and depending on a current xerces version.

Then, when testing SVG emission from darkstar,
found out that my spec is being parsed correctly
or overcoming defeciencies in the multi-layer plot
where we have some layers overriding a shared y component.

This is problematic because apparently it janks with the
underlying data since it's working on the literal
json data object directly (eg. mutating).

So we have a problem that vega 5.2 is actually
cleaning up for us.  When we go to use darkstar to
expore svg, we get errors during validation since
a whole bunch of weird NANs and MNaNs show up, despite
the spec rendering correctly in the browser.

Some thoughts: we can try to bump the darkstar vega impl
to 5.2 (which is what Oz is using), and see if that works out.

Just copy the assets from roz over and see if we get a quick win...

We can try to find the Proper Way to handle layered plots and
dodge the warning too.

Still wrapping up the nlist export, but we're pretty close.